+++ fixtures/flight-browser/index.html
@@ -19,8 +19,8 @@
     <script src="../../build/dist/react.development.js"></script>
     <script src="../../build/dist/react-dom.development.js"></script>
     <script src="../../build/dist/react-dom-server.browser.development.js"></script>
-    <script src="../../build/dist/react-dom-unstable-flight-server.browser.development.js"></script>
-    <script src="../../build/dist/react-dom-unstable-flight-client.development.js"></script>
+    <script src="../../build/dist/react-flight-dom-webpack-server.browser.development.js"></script>
+    <script src="../../build/dist/react-flight-dom-webpack.development.js"></script>
     <script src="https://unpkg.com/babel-standalone@6/babel.js"></script>
     <script type="text/babel">
       let Suspense = React.Suspense;
