--- packages/react-dom/npm/unstable-flight-client.js
@@ -1,7 +1,0 @@
-'use strict';
-
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-dom-unstable-flight-client.production.min.js');
-} else {
-  module.exports = require('./cjs/react-dom-unstable-flight-client.development.js');
-}
