+++ packages/react-flight-dom-webpack/npm/server.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('./server.node');
