+++ packages/react-flight-dom-webpack/server.js
@@ -9,4 +9,4 @@
 
 'use strict';
 
-module.exports = require('./unstable-flight-server.node');
+module.exports = require('./server.node');
