+++ packages/react-flight-dom-webpack/src/__tests__/ReactFlightDOMBrowser-test.js
@@ -12,39 +12,20 @@
 
 // Polyfills for test environment
 global.ReadableStream = require('@mattiasbuelens/web-streams-polyfill/ponyfill/es6').ReadableStream;
+global.TextEncoder = require('util').TextEncoder;
 global.TextDecoder = require('util').TextDecoder;
 
-let Stream;
 let React;
 let ReactFlightDOMServer;
 let ReactFlightDOMClient;
 
-describe('ReactFlightDOM', () => {
+describe('ReactFlightDOMBrowser', () => {
   beforeEach(() => {
     jest.resetModules();
-    Stream = require('stream');
     React = require('react');
-    ReactFlightDOMServer = require('react-dom/unstable-flight-server');
-    ReactFlightDOMClient = require('react-dom/unstable-flight-client');
+    ReactFlightDOMServer = require('react-flight-dom-webpack/server.browser');
+    ReactFlightDOMClient = require('react-flight-dom-webpack');
   });
-
-  function getTestStream() {
-    let writable = new Stream.PassThrough();
-    let readable = new ReadableStream({
-      start(controller) {
-        writable.on('data', chunk => {
-          controller.enqueue(chunk);
-        });
-        writable.on('end', () => {
-          controller.close();
-        });
-      },
-    });
-    return {
-      writable,
-      readable,
-    };
-  }
 
   async function waitForSuspense(fn) {
     while (true) {
@@ -60,7 +41,7 @@ describe('ReactFlightDOM', () => {
     }
   }
 
-  it('should resolve HTML using Node streams', async () => {
+  it('should resolve HTML using W3C streams', async () => {
     function Text({children}) {
       return <span>{children}</span>;
     }
@@ -80,9 +61,8 @@ describe('ReactFlightDOM', () => {
       return model;
     }
 
-    let {writable, readable} = getTestStream();
-    ReactFlightDOMServer.pipeToNodeWritable(<App />, writable);
-    let result = ReactFlightDOMClient.readFromReadableStream(readable);
+    let stream = ReactFlightDOMServer.renderToReadableStream(<App />);
+    let result = ReactFlightDOMClient.readFromReadableStream(stream);
     await waitForSuspense(() => {
       expect(result.model).toEqual({
         html: '<div><span>hello</span><span>world</span></div>',
