+++ packages/react-server/src/ReactDOMServerFormatConfig.js
